+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Change log
 - Added a ``LanguageAndTranslationOfMixin`` which not only allows
   defining the language of objects but also defining objects to be
   translations of other objects.
+- Added Travis CI jobs for Django 3.1b1 and Python 3.8.
 
 
 `0.34`_ (2020-06-05)