+++ CHANGELOG.rst
@@ -6,12 +6,17 @@ Change log
 `Next version`_
 ~~~~~~~~~~~~~~~
 
+
+`0.35`_ (2020-07-28)
+~~~~~~~~~~~~~~~~~~~~
+
 - **(not yet) BACKWARDS INCOMPATIBLE** Moved the ``feincms3.apps``
   module to :mod:`feincms3.applications`. The reason for this change is
   that Django 3.2 will start autodiscovering app configs and therefore
   automatically loads the ``.apps`` submodule of all entries in
   ``INSTALLED_APPS``. This leads to a crash when the ``.apps`` module
-  contains models (such as our ``AppsMixin``).
+  contains models (such as our ``AppsMixin``). ``feincms3.apps`` isn't
+  populated from Django 3.2 upwards because of this.
 - Fixed an infinite recursion crash when referencing pages using
   ``on_delete=SET_NULL``
 - Added a ``LanguageAndTranslationOfMixin`` which not only allows
@@ -617,4 +622,5 @@ functionality.
 .. _0.32: https://github.com/matthiask/feincms3/compare/0.31...0.32
 .. _0.33: https://github.com/matthiask/feincms3/compare/0.32...0.33
 .. _0.34: https://github.com/matthiask/feincms3/compare/0.33...0.34
-.. _Next version: https://github.com/matthiask/feincms3/compare/0.34...main
+.. _0.35: https://github.com/matthiask/feincms3/compare/0.34...0.35
+.. _Next version: https://github.com/matthiask/feincms3/compare/0.35...main